requirements.txt
django==3.2
djangorestframework
drf-yasg
django-filter
django-cors-headers
psycopg2-binary
gunicorn
telepot
